Use size_t for new crypto size parameters.

+++ crypto/dh/dh_pmeth.c
@@ -187,7 +187,7 @@ static int pkey_dh_keygen(EVP_PKEY_CTX *ctx, EVP_PKEY *pkey)
 	return DH_generate_key(pkey->pkey.dh);
 	}
 
-static int pkey_dh_derive(EVP_PKEY_CTX *ctx, unsigned char *key, int *keylen)
+static int pkey_dh_derive(EVP_PKEY_CTX *ctx, unsigned char *key, size_t *keylen)
 	{
 	int ret;
 	if (!ctx->pkey || !ctx->peerkey)
